Move HTML::Pipeline filters under Dradis:: namespace
